chore: move template installing functions to utils
