+++ ui/src/templates/utils/index.ts
@@ -8,6 +8,12 @@ import {
   Variable,
 } from 'src/types'
 
+import {
+  Error as PkgError,
+  postTemplatesApply,
+  TemplateSummary,
+} from 'src/client'
+
 export function findIncludedsFromRelationships<
   T extends {id: string; type: TemplateType}
 >(
@@ -89,4 +95,38 @@ export const getTemplateNameFromGithubUrl = (url: string): string => {
 
 export const getGithubUrlFromTemplateName = (templateName: string): string => {
   return `https://github.com/influxdata/community-templates/tree/master/${templateName}`
+}
+
+const applyTemplates = async (params) => {
+  const resp = await postTemplatesApply(params)
+  if (resp.status >= 300) {
+    throw new Error((resp.data as PkgError).message)
+  }
+
+  const summary = (resp.data as TemplateSummary).summary
+  return summary
+}
+
+export const reviewTemplate = async (orgID: string, templateUrl: string) => {
+  const params = {
+    data: {
+      dryRun: true,
+      orgID,
+      remotes: [{url: templateUrl}],
+    },
+  }
+
+  return applyTemplates(params)
+}
+
+export const installTemplate = async (orgID: string, templateUrl: string) => {
+  const params = {
+    data: {
+      dryRun: false,
+      orgID,
+      remotes: [{url: templateUrl}],
+    },
+  }
+
+  return applyTemplates(params)
 }
